TCL: Code relocation in HttpCompact and HttpBodyFactory for TCL removal
