+++ .gitignore
@@ -97,7 +97,7 @@ iocore/eventsystem/test_MIOBufferWriter
 iocore/hostdb/test_RefCountCache
 
 proxy/hdrs/test_mime
-proxy/http/test_ForwardedConfig
+proxy/http/test_proxy_http
 proxy/http2/test_Huffmancode
 proxy/http2/test_Http2DependencyTree
 proxy/http2/test_HPACK
