+++ proxy/http/Makefile.am
@@ -81,22 +81,32 @@ libhttp_a_SOURCES += HttpUpdateTester.cc \
 	RegressionHttpTransact.cc
 endif
 
-check_PROGRAMS = \
-test_ForwardedConfig
+check_PROGRAMS = test_proxy_http
 
 TESTS = $(check_PROGRAMS)
 
-test_ForwardedConfig_CPPFLAGS = $(AM_CPPFLAGS)\
+test_proxy_http_CPPFLAGS = $(AM_CPPFLAGS)\
 	-I$(abs_top_srcdir)/tests/include
 
-test_ForwardedConfig_SOURCES = \
+test_proxy_http_SOURCES = \
+	unit_tests/unit_test_main.cc \
 	unit_tests/test_ForwardedConfig.cc \
 	ForwardedConfig.cc \
-	unit_tests/test_ForwardedConfig_mocks.cc
+	unit_tests/test_error_page_selection.cc \
+	HttpBodyFactory.cc \
+	HttpBodyFactory.h
 
-test_ForwardedConfig_LDADD = \
+test_proxy_http_LDADD = \
 	$(top_builddir)/src/tscore/libtscore.la \
-	$(top_builddir)/src/tscpp/util/libtscpputil.la
+	$(top_builddir)/src/tscpp/util/libtscpputil.la \
+	$(top_builddir)/proxy/hdrs/libhdrs.a \
+	$(top_builddir)/iocore/eventsystem/libinkevent.a \
+	$(top_builddir)/lib/records/librecords_p.a \
+	$(top_builddir)/proxy/shared/libUglyLogStubs.a \
+	$(top_builddir)/proxy/logging/liblogging.a \
+	$(top_builddir)/mgmt/libmgmt_p.la \
+	@HWLOC_LIBS@ \
+	@LIBTCL@ @LIBCAP@
 
 clang-tidy-local: $(libhttp_a_SOURCES) $(noinst_HEADERS)
 	$(CXX_Clang_Tidy)
